Fix for unaffected transparency
When the mouse enters or leave the image window, the opacity should change, but it is unaffected. Fixed after this commit.
Renamed and refactored opacity-related properties and event.
Remove some redundant code.
Note: Due to the modification of properties names, after this commit, the old configuration file "SetunaConfig.xml" needs to be deleted.